requirements.txt
python-dotenv==1.0.1
google-genai==1.65.0
pydantic==2.12.0
